Game Tidy Up#1
Bug fixes, added a flag instead of justing changing the materials of the
flagged tile. There is also a skybox now

Bug#1: It was possible to flag and unflag a tile, even when the game was
paused

Bug#2: It was also possible to go -1 in the flag amount and farm them,
so instead of just having 15, you could have 16 17 18 and so on.